Module astro::angle

Angles for astronomy

Functions

angl_sepr
Returns the angular separation between two angular points

deg_frm_dms
Returns an angle in degrees with decimals, from an angle expressed in degrees, arcminutes and arcseconds

deg_frm_hms
Returns an angle in degrees with decimals, from an angle expressed in hours, minutes and seconds

dms_frm_deg
Returns an angle expressed in degrees, arcminutes and arcseconds, from an angle in degrees with decimals

hms_frm_deg
Returns an angle expressed in hours, minutes and seconds, from an angle in degrees with decimals

limit_to_360
Returns the equivalent angle in [0, 360] degree range