Module astro::atmos

Corrections for atmospheric refraction

Functions

refrac_by_pressr
Computes the refraction term modifier for local pressure

refrac_by_temp
Computes the refraction term modifier for local temperature

refrac_frm_apprnt_alt
Computes the refraction term for true altitude

refrac_frm_apprnt_alt_15
Computes the refraction term for true altitudes greater than 15 degrees

refrac_frm_true_alt
Computes the refraction term for apparent altitude

refrac_frm_true_alt_15
Computes the refraction term for apparent altitudes greater than 15 degrees