Module astro::coords

Transform between coordinate systems

Structs

EclPoint
Represents a point in the ecliptic coordinate system

EqPoint
Represents a point in the equatorial coordinate system

GeographPoint
Represents a point on the geographical surface of the Earth

Functions

alt_frm_eq
Returns the altitude from equatorial coordinates

asc_frm_ecl
Returns the right ascension from ecliptic coordinates

asc_frm_gal
Returns the right ascension from galactic coordinates

az_frm_eq
Returns the azimuth from equatorial coordinates

dec_frm_ecl
Returns the declination from ecliptic coordinates

dec_frm_gal
Returns the declination from galactic coordinates

dec_frm_hz
Returns the declination from local horizontal coordinates

ecl_lat_frm_eq
Returns the ecliptic latitude from equatorial coordinates

ecl_long_frm_eq
Returns the ecliptic longitude from equatorial coordinates

gal_lat_frm_eq
Returns the galactic latitude from equatorial coordinates

gal_long_frm_eq
Returns the galactic longitude from equatorial coordinates

hr_angl_frm_hz
Returns the hour angle from local horizontal coordinates

hr_angl_frm_loc_sidr
Returns the hour angle from local sidereal time and right ascension

hr_angl_frm_observer_long
Returns the hour angle from geographical longitude and Greenwhich sidereal time