Module astro::ecliptic

The Earth's ecliptic

Functions

angl_betwn_eclip_and_hz
Computes the angle between the ecliptic and a horizon on Earth

eclip_points_on_hz
Computes the longitudes of the two ecliptic points on a horizon on Earth

mn_oblq_IAU
Computes the mean obliquity of the ecliptic using the IAU formula

mn_oblq_laskar
Computes the mean obliquity of the ecliptic using J. Laskar's formula