Module astro::lunar

The Earth's moon

Enums

Phase
Represents a phase of the Moon

Functions

bright_limb
Returns the position angle of the Moon's bright limb

eq_hz_parllx
Returns the equatorial horizontal parallax of the Moon

geocen_ecl_pos
Returns the geocentric ecliptic position of the Moon, referred to the mean equinox of the date

illum_frac_frm_ecl_coords
Returns the illuminated fraction of the lunar disk, using eclipctic coordinates

illum_frac_frm_eq_coords
Returns the illuminated fraction of the lunar disk, using equatorial coordinates

inc_of_mn_lunar_eq
Returns the inclination of the mean lunar equator with the ecliptic

mn_ascend_node
Returns the longitude of the mean ascending node of the Moon

mn_perigee
Returns the longitude of the mean perigee of the Moon

optical_libr
Returns the optical librations of the Moon in longitude and latitude

physical_libr
Returns the physical librations of the Moon in longitude and latitude

pos_angl_of_axis_of_rot
Returns the position angle of the axis of rotation of the Moon

semidiameter
Returns the equatorial semidiameter of the Moon

time_of_passage_through_nodes
Returns the times of passage of the Moon through the ascending and descending nodes, close to a given date

time_of_phase
Returns the Julian day corresponding to one of the four phases of the Moon

topocen_libr_by_diff_corrections
Returns the topocentric librations of the Moon

total_libr
Returns the total librations of the Moon in longitude and latitude

true_ascend_node
Returns the longitude of the true ascending node of the Moon