Function astro::lunar::pos_angl_of_axis_of_rot

pub fn pos_angl_of_axis_of_rot(JD: f64, mn_ascent_node_long: f64, total_lib_lat: f64, nut_in_long: f64, true_oblq: f64, apprnt_moon_asc: f64) -> f64

Computes the position angle of the axis of rotation of the Moon

Returns

pos_angl_of_axis_of_rot: The position angle of the axis of rotation of the Moon | in radians

Arguments

JD : Julian (Ephemeris) day

mn_ascent_node_long: Longitude of the mean ascending node of the Moon | in radians

total_lib_lat : Total libration of the Moon in latitude | in radians

nut_in_long : Nutation correction for longitude | in radians

true_oblq : True obliquity of the ecliptic | in radians

apprnt_moon_asc : Apparent geocentric right ascension of the Moon | in radians