Function astro::lunar::time_of_phase

pub fn time_of_phase(date: &Date, phase: &Phase) -> f64

Computes the Julian day corresponding to one of the four phases of the Moon

Returns

JD: Julian day corresponding to the exact time of the phase that is closest to date

Meeus says the mean error in JD for phases between 1980 AD and mid-2020 AD is 3.8 seconds.

Arguments

date : Date of interest, close to the phase

phase: The Phase