Function astro::lunar::topocent_libr_by_diff_corrections

pub fn topocent_libr_by_diff_corrections(observer_lat: f64, geocent_dec_moon: f64, local_hour_angl: f64, geocent_hz_parllx_moon: f64, pos_angl_axis_of_rot: f64, total_lib_lat: f64) -> (f64, f64, f64)

Computes the topocentric librations of the Moon

Returns

(topocent_libr_in_long, topocent_libr_in_lat, topocent_libr_in_P)

topocent_libr_in_long: Topocentric libration in longitude | in radians

topocent_libr_in_lat : Topocentric libration in latitude | in radians

topocent_libr_in_P : Topocentric libration in position angle of the axis of rotation | in radians

Arguments

observer_lat : Latitude of the observer | in radians

geocent_dec_moon : Geocentric decation of the Moon | in radians

local_hour_angl : Local hour angle of the Moon | in radians

geocent_hz_parllx_moon: Geocentric equatorial horizontal parallax of the Moon | in radians

pos_angl_axis_of_rot : Position angle of the axis of rotation of the Moon | in radians

total_lib_lat : Total libration of the Moon in latitude | in radians