Function astro::misc::parllc_angl_on_hz

pub fn parllc_angl_on_hz(observer_lat: f64, dec: f64) -> f64

Computes the parallactic angle of a celestial body on the horizon

Returns

parallac_angle: Parallactic angle of the celestial body on the horizon | in radians

Arguments

observer_lat: The observer's geographical latitude | in radians

dec : Declination of the celestial body | in radians