Function astro::nutation::nutation_in_eq_coords

pub fn nutation_in_eq_coords(eq_point: &EqPoint, nut_in_long: f64, nut_in_oblq: f64, tru_oblq: f64) -> (f64, f64)

Computes nutation in equatorial coordinates

Returns

(nut_in_asc, nut_in_dec)

nut_in_asc: Nutation in right ascension | in radians

nut_in_dec: Nutation in declination | in radians

Arguments

eq_point : Equatorial point uncorrected for nutation | in radians

nut_in_long: Nutation in longitude | in radians

nut_in_oblq: Nutation in obliquity | in radians

tru_oblq : True obliquity of the ecliptic | in radians

The declination of eq_point should not be close to either of the two of the celestial poles, as the values of nutation returned here are only from first-order corrections.