Module astro::orbit::elliptic

Elliptic orbits

Functions

aph_vel
Computes the velocity of a body at aphelion in an elliptic orbit

ecc_anom
Computes the eccentric anomaly of a body in an elliptic orbit

length
Computes the approximate length of an ellipse

length_ramanujan
Computes the approximate length of an ellipse using the Ramanujan method

mn_motion
Computes the mean motion of an elliptic orbit

passage_through_node
Computes the time of passage of a body through a node of an elliptic orbit, and it's radius vector at that time

perih_vel
Computes the velocity of a body at perihelion in an elliptic orbit

rad_vec_frm_ecc_anom
Computes the radius vector of a body in an elliptic orbit from it's eccentric anomaly

rad_vec_frm_true_anom
Computes the radius vector of a body in an elliptic orbit from it's true anomaly

semimaj_axis
Computes the semimajor axis of an elliptic orbit

true_anom
Computes the true anomaly of a body in an elliptic orbit

vel
Computes the velocity of a body in an elliptic orbit