Function astro::orbit::elliptic::length_ramanujan

pub fn length_ramanujan(a: f64, b: f64) -> f64

Computes the approximate length of an ellipse using the Ramanujan method

Returns

approx_length: An approximate value for the length of the ellipse (same unit as that of a and b), using a formula given by Ramanujan in 1914.

The error in approx_length is:

0% for a = b

0.4155% for e = 1

Arguments

a: Semimajor axis of the ellipse (same unit as that of b)

b: Semiminor axis of the ellipse (same unit as that of a)