Function astro::orbit::elliptic::passage_through_node

pub fn passage_through_node(w: f64, n: f64, a: f64, e: f64, T: f64, node: &Node) -> (f64, f64)

Computes the time of passage of a body through a node of an elliptic orbit, and it's radius vector at that time

Returns

(time_of_pass, rad_vec)

time_of_pass: Time of passage through the node, in Julian (Ephemeris) day

rad_vec : Radius vector of the body at the time of passage | in AU

Arguments

w : Argument of the perihelion | in radians

n : Mean motion of the orbit (radians per day)

a : Semimajor axis of the orbit | in AU

e : Eccentricity of the orbit

T : Time of passage in perihelion, in Julian (Ephemeris) day

node: Ascend or Descend node