Function astro::orbit::near_parabolic::true_anom_and_rad_vec

pub fn true_anom_and_rad_vec<'a>(t: f64, T: f64, ecc: f64, q: f64, accuracy: f64) -> Result<(f64, f64), &'a str>

Computes the true anomaly and radius vector of a body in a near-parabolic orbit at a given time

Returns

(true_anom, rad_vec)

true_anom: True anomaly of the body at time t | in radians

rad_vec : Radius vector of the body at time t | in AU

Arguments

t : Time of interest, in Julian (Ephemeris) day

T : Time of passage through the perihelion, in Julian (Ephemeris) day

ecc : Eccentricity of the near-parabolic orbit

q : Perihelion distance | in AU

accuracy: Desired accuracy for the results. For example, passing 0.000001 gives that much accuracy in radians and AU.