Module astro::planet

8 Planets of the Solar System

Modules

earth
The Earth

jupiter
Jupiter

mars
Mars

saturn
Saturn

Enums

Planet
Represents a planet

Functions

apprnt_mag_84
Computes a planet's apparent magnitude using the Astronomical Almanac's method adopted in 1984

apprnt_mag_muller
Computes a planet's apparent magnitude using G. Muller's formulae

ecl_coords_to_FK5
Computes a planet's geocentric ecliptic coordinates converted to the FK5 system

geocent_apprnt_ecl_coords
Computes a planet's geocentric, apparent ecliptic position, corrected for light-time

geocent_eq_coords

geocent_geomet_ecl_coords
Computes a planet's geocentric, geometric ecliptic position, uncorrected for light-time

heliocent_coords
Computes a planet's heliocentric coordinates, referred to the mean equinox of the date

heliocent_coords_frm_orb_elements

illum_frac_frm_dist
Computes the illuminated fraction of a planet's disk from it's distance to the Sun and Earth

illum_frac_frm_phase_angl
Computes the illuminated fraction of a planet's disk from it's phase angle

orb_elements
Computes a planet's orbital elements, referred to the mean equinox of the date

phase_angl
Computes a planet's phase angle

pos_angle_of_bright_limb
Computes the position angle of the bright limb of a planet

semidiameter
Computes a planet's equatorial semidiameter