Module astro::planet::earth

The Earth

Functions

angl_betwn_diurnal_path_and_hz
Computes the angle between diurnal path and the horizon

approx_geodesic_dist
Computes a low accuracy geodesic distance between two points on the Earth's surface | in kilometers

ecc_of_meridian
Returns the eccentricity of the Earth's meridian

eq_rad
Returns the equatorial radius of the Earth | in kilometers

equation_of_time
Computes the equation of time | in radians

flat_fac
Returns the flattening factor of the Earth

geodesic_dist
Computes a high accuracy geodesic distance between two points on the Earth's surface | in kilometers

geograph_geocent_lat_diff
Computes the difference between the geographic latitude and geocentric latitude

linear_velocity_at_lat
Computes the linear velocity of a point at a latitude

pol_rad
Returns the polar radius of the Earth | in kilometers

rad_curv_of_meridian
Computes the radius of curvature of the Earth's meridian at a latitude

rad_of_parll_lat
Computes the radius of the parallel of a latitude

rho
Computes the distance from the Earth's center to a point on the ellipsoid

rho_sin_cos_phi
Computes two quantities that are used elsewhere in the library

rot_angular_velocity
Returns the rotational angular velocity of the Earth | in radian per second