Function astro::planet::jupiter::moon::apprnt_rect_coords

pub fn apprnt_rect_coords(JD: f64, moon: &Moon) -> (f64, f64)

Computes the apparent rectangular coordinates for a Galilean moon

This function implements the low accuracy method described in Meeus's book, that is "sufficient for identifying the satellites at the telescope, or drawing a wavy-line diagram showing their positions with respect to Jupiter"

Returns

(X, Y)

The rectangular coordinates returned give the apparent position of a moon, with respect to Jupiter, as seen from Earth. The X and Y coordinates are measured from the center of the disk of Jupiter, in units of Jupiter's equatorial radius.

X is measured positively to the west of Jupiter, and negatively to the east. The x-axis coincides with Jupiter's equator.

Y is measured positively to the north of Jupiter, and negatively to the south. The y-axis coincides with Jupiter's axis of rotation.

Arguments

JD : Julian (Ephemeris) day

moon: The Moon