Struct astro::planet::jupiter::Ephemeris

pub struct Ephemeris {
    pub De: f64,
    pub Ds: f64,
    pub P: f64,
    pub w1: f64,
    pub w2: f64,
}

Holds Jupiter's ephemeris values for physical observations

Fields

De

Jupiter-centric declination of the Earth

Ds

Jupiter-centric declination of the Sun

P

Geocentric position angle of Jupiter's northern rotation pole, or also called, position angle of the axis

w1

Longitude of the central meridian for Rotational System I

w2

Longitude of the central meridian for Rotational System II