Module astro::planet::mars

Mars

Structs

Ephemeris
Holds Mar's ephemeris values for physical observations

Functions

ephemeris
Computes quantites used in the ephemeris for physical observations of Mars

north_pol_ecl_coords
Computes the ecliptic coordinates of Mars's north pole, referred to the mean equinox of the date

north_pol_eq_coords_J1950
Returns the equatorial coordinates of Mars's north pole for the epoch J1950.0

north_pol_eq_coords_J2000
Returns the equatorial coordinates of Mars's north pole for the epoch J2000.0