Struct astro::planet::saturn::ring::Elements

pub struct Elements {
    pub B: f64,
    pub B1: f64,
    pub P: f64,
    pub deltaU: f64,
    pub a: f64,
    pub b: f64,
}

Holds the elements for the ring system of Saturn

Fields

B

Saturnicentric latitude of the Earth, referred to the plane of the ring

B1

Saturnicentric latitude of the Sun, referred to the plane of the ring

P

Geocentric position angle of the northern semiminor axis of the apparent ellipse of the ring of the axis

deltaU

Difference between Saturnicentric longitudes of the Sun and the Earth, measured in the plane of the ring

a

Major axis of the outer edge of the outer ring

b

Minor axis of the outer edge of the outer ring