Module astro::precess

Corrections for precession

Functions

annual_precess
Computes annual precession in equatorial coordinates towards a new epoch

precess_ecl_coords
Computes ecliptic coordinates reduced to a different epoch

precess_eq_coords
Computes equatorial coordinates reduced to a different epoch

precess_eq_coords_FK5
Computes equatorial coordinates, from coordinates referred to the FK4 system, reduced to a different epoch

precess_orb_elements
Computes orbital elements reduced to a different equinox