Function astro::precess::annual_precess

pub fn annual_precess(asc: f64, dec: f64, JD: f64) -> (f64, f64)

Computes annual precession in equatorial coordinates towards a new epoch

Returns

(ann_precess_asc, ann_precess_dec)

ann_precess_asc: Annual precession in right ascension towards the new epoch | in radians

ann_precess_dec: Annual precession in declination towards the new epoch | in radians

In the case of a star, the precession returned here does not take into account it's annual proper motion. The annual proper motion in right ascension and declination must simply be added to the corresponding annual precession returned here in order to reduce the equatorial coordinates to the new epoch.

Arguments

asc: Right ascension for the old epoch | in radians

dec: Declination for the old epoch | in radians. Shouldn't be too close to the celestial poles of the Earth.

JD : Julian (Ephemeris) day corresponding to the new epoch. Shouldn't be more than a few hundred years away from the old epoch.