Function astro::precess::precess_orb_elements

pub fn precess_orb_elements(old_inc: f64, old_arg_perih: f64, old_long_ascend_node: f64, JD1: f64, JD2: f64) -> (f64, f64, f64)

Computes orbital elements reduced to a different equinox

Returns

(new_inc, new_arg_perih, new_long_ascend_node)

new_inc : Inclination in the new equinox | in radians

new_arg_perih : Argument of perihelion in the new equinox | in radians

new_long_ascend_node: Longitude of ascending node in the new equinox | in radians

Arguments

old_inc : Inclination in the old equinox | in radians

old_arg_perih : Argument of perihelion in the old equinox | in radians

old_long_ascend_node: Longitude of ascending node in the old equinox | in radians

JD1 : Julian (Ephemeris) day corresponding to the old equinox

JD2 : Julian (Ephemeris) day corresponding to the new equinox