Module astro::star

Stars

Functions

abs_mag_frm_dist
Computes the absolute magnitude of a star from its distance from earth

abs_mag_frm_parallax
Computes the absolute magnitude of a star from its parallax

angl_between_north_celes_and_eclip_pole
Computes the angle between a vector from a star to the north celestial pole of the Earth and a vector from the same star to the north pole of the ecliptic

brightness_ratio
Computes the brightness ratio of two stars

combined_mag
Computes the combined magnitude of two stars

combined_mag_of_many
Computes the combined magnitude of two or more stars

eq_coords_frm_motion
Computes the equatorial coordinates of a star at at a different time from it's motion in space

mag_diff
Computes the difference in magnitude of two stars

proper_motion_in_eq_coords