Function astro::star::eq_coords_frm_motion

pub fn eq_coords_frm_motion(asc0: f64, dec0: f64, r: f64, delta_r: f64, proper_motion_asc: f64, proper_motion_dec: f64, t: f64) -> (f64, f64)

Computes the equatorial coordinates of a star at at a different time from it's motion in space

This function Computes the equatorial coordinates of a star at a different time by taking into account it's proper motion, distance and radial velocity.

Returns

(new_asc, new_dec)

new_asc: Right ascension at the different time | in radians

new_dec: Declination at the different time | in radians

Arguments

asc0: Right ascension of the star initially | in radians

dec0: Declination of the star initially | in radians

r: Distance of the star (parsecs)

delta_r: Radial velocity of the star (parsecs/second)

proper_motion_asc: Proper motion of the star in right ascension | in radians

proper_motion_dec: Proper motion of the star in declination | in radians

t: Decimal years from the inital time; negative in the past and positive in the future