Module astro::sun

The Sun

Functions

ecl_coords_to_FK5
Computes the Sun's geocentric ecliptic coordinates converted to the FK5 system

ephemeris
Return quantites used in the ephemeris for physical observations of the Sun

geocent_ecl_pos
Computes the Sun's geocentric ecliptic position, referred to the mean equinox of the date

geocent_rect_coords
Computes the Sun's geocentric rectangular coordinates, referred to the mean equinox of the date

semidiameter
Computes the Sun's equatorial semidiameter

synodic_rot
Computes an approximate time for the beginning of a solar synodic rotation