Function astro::sun::ecl_coords_to_FK5

pub fn ecl_coords_to_FK5(JD: f64, ecl_long: f64, ecl_lat: f64) -> (f64, f64)

Computes the Sun's geocentric ecliptic coordinates converted to the FK5 system

Returns

(ecl_long_FK5, ecl_lat_FK5)

ecl_long_FK5: Ecliptic longitude of the Sun | in radians, converted to the FK5 system

ecl_lat_FK5 : Ecliptic latitude of the Sun | in radians, converted to the FK5 system

Arguments

JD : Julian (Ephemeris) day

ecl_long: Ecliptic longitude of the Sun on JD | in radians, referred to the mean equinox of the date

ecl_lat : Ecliptic latitude of the Sun JD | in radians, referred to the mean equinox of the date