Function astro::sun::synodic_rot

pub fn synodic_rot(C: i64) -> f64

Computes an approximate time for the beginning of a solar synodic rotation

Returns

JD: The Julian Day corresponding to the approximate time for the beginning of a solar synodic rotation

Between the years 1850 and 2100, JD will be in less than 0.002 days in error.

Arguments

C: Carrington's synodic rotation number