Function astro::transit::time

pub fn time(transit_type: &TransitType, transit_body: &TransitBody, geograph_point: &GeographPoint, eq_point1: &EqPoint, eq_point2: &EqPoint, eq_point3: &EqPoint, apprnt_greenwhich_sidr: f64, delta_t: f64, moon_eq_hz_parallax: f64) -> (i64, i64, f64)

Computes the time of transit for a celestial body

Returns

(hour, min, sec): Time of transit on the day of interest, in UTC

Arguments

transit_type : A TransitType

transit_body : The TransitBody

geograph_point: Geographic point of the observer | in radians

Let JD be the Julian (Ephemeris) day of interest,

eq_point1: Equatorial point of the transit body on JD - 1 | in radians

eq_point2: Equatorial point of the transit body on JD | in radians

eq_point3: Equatorial point of the transit body on JD + 1 | in radians

apprnt_greenwhich_sidr: Apparent sidereal time at Greenwhich on JD | in radians

delta_t: ΔT for JD (Julian day)

moon_eq_hz_parallax: Equatorial horizontal parallax of the Moon on JD | in radians. Pass a meaningfull value here only when TransitBody::Moon is passed for transit_body.