Module astro::angle

Angles for astronomy

Constants

TWO_PI

Functions

anglr_sepr
Computes the angular separation between two angular points

deg_frm_dms
Computes an angle in degrees with decimals, from an angle expressed in degrees, arcminutes and arcseconds

deg_frm_hms
Computes an angle in degrees with decimals, from an angle expressed in hours, minutes and seconds

dms_frm_deg
Computes an angle expressed in degrees, arcminutes and arcseconds, from an angle in degrees with decimals

hms_frm_deg
Computes an angle expressed in hours, minutes and seconds, from an angle in degrees with decimals

limit_to_360
Computes the equivalent angle in [0, 360] degree range

limit_to_two_PI
Computes the equivalent angle in [0, 2π] radian range