Module astro::lunar

The Earth's moon

Enums

Phase
Represents a phase of the Moon

Functions

bright_limb
Computes the position angle of the Moon's bright limb

eq_hz_parllx
Computes the equatorial horizontal parallax of the Moon

geocent_ecl_pos
Computes the geocentric ecliptic position of the Moon, referred to the mean equinox of the date

illum_frac_frm_ecl_coords
Computes the illuminated fraction of the lunar disk, using eclipctic coordinates

illum_frac_frm_eq_coords
Computes the illuminated fraction of the lunar disk, using equatorial coordinates

inc_of_mn_lunar_eq
Computes the inclination of the mean lunar equator with the ecliptic

mn_ascend_node
Computes the longitude of the mean ascending node of the Moon

mn_perigee
Computes the longitude of the mean perigee of the Moon

optical_libr
Computes the optical librations of the Moon in longitude and latitude

physical_libr
Computes the physical librations of the Moon in longitude and latitude

pos_angl_of_axis_of_rot
Computes the position angle of the axis of rotation of the Moon

semidiameter
Computes the equatorial semidiameter of the Moon

time_of_passage_through_nodes
Computes the times of passage of the Moon through the ascending and descending nodes, close to a given date

time_of_phase
Computes the Julian day corresponding to one of the four phases of the Moon

topocent_libr_by_diff_corrections
Computes the topocentric librations of the Moon

total_libr
Computes the total librations of the Moon in longitude and latitude

true_ascend_node
Computes the longitude of the true ascending node of the Moon