Function astro::planet::orb_elements

pub fn orb_elements(planet: &Planet, JD: f64) -> (f64, f64, f64, f64, f64, f64, f64, f64)

Computes a planet's orbital elements, referred to the mean equinox of the date

Returns

(L, a, e, i, omega, pi, M, w)

L: Mean longitude | in radians

a: Semimajor axis of the orbit | in AU

e: Eccentricity of the orbit

i: Inclination of the plane of the orbit with the plane of the Earth's ecliptic | in radians

omega: Longitude of the ascending node | in radians. An undefined value is returned for Planet::Earth.

pi: Longitude of the perihelion | in radians

M: Mean anomaly | in radians

w: Argument of the perihelion | in radians. An undefined value is returned for Planet::Earth.

Arguments

planet: Any variant of Planet

JD : Julian (Ephemeris) day